requirements.txt
fastapi==0.135.3
uvicorn==0.42.0
pydantic==2.12.5
openai==2.30.0
python-dotenv==1.2.2
openenv-core==0.2.3